requirements.txt
Flask
flask_cors
google-generativeai
gunicorn